Case sensitive file system support for AnimViewer.
With this patch the case of each part of the path or file name being loaded/saved is checked against the file system, so if all your files are lowercase or mixed case it will still work.
E.g. data instead of DATA.
On case insensitive file systems nothing additional is done.

Fixes issue 1136.
Thanks, Stephen!