This patch widens the space for custom game names a little.
It also fixes a contradiction - at the start of the demo level it tells you "Unfortunately the demo version only contains this level (apart from custom levels)." 

If you are meant to be able to play custom maps with demo data, then perhaps we need to remove the notice and allow the game to load them. Removing the contradiction seems to be the best option at the moment.
This can be discussed (issue 1454).

Thanks, Mark!