Key presses
To speed up the game press Z, on release the speed changes to what it
was.

Moved transparent walls from a debug option to the main game. Press X
and the walls become transparent and return to normal on release of the
key.
Whilst the walls are transparent you cannot buy plots of land.
this should fix issue 1470
Issue 1493 and Issue 1468 are also related