Issue #90
This is a possible fix for issue #90.
I have moved he message about the missing graphic from research screen
to bottom panel as I believe it was failing due to the extras that are
in the open research screen button function compared to the other
buttons.  It should now act as intended and open the no graphics info is
demo data is used and should no longer have the advisor give incorrect
information about needing to build the already build research room -
demo again.