Let the game respect "no earthquakes"
Small change that should let the game respect the mapmakers decision to
not have earthquakes in a level by setting the severity to 0.
Unsure if this will also stop random quakes in games were there are no
pre-determined quakes left in the level file as I need a long running
save.