Remember players choice
small change that will allow the game to save the players choice of heat
display colours.
Similar change is needed for the grant/deny option as currently it can
only be changed by editing the config if you want it to stick -
otherwise it reverts back to the config setting and your choice is not
saved.