Add documentation about issue templates
Mailed the GitHub guys about this and they said that it is currently not possible to let bug reporters fill this in automaticly, as you could with Google Code, but they will might do this in the future. So, this issue is a kind of workaround till GitHub fully supports issue templates.
They said the following:

"Adding an issue template is not possible currently, but we might add it in the future. Thanks for the suggestion, I'll add it to the Feature Request List so the team can see it.

In the meantime, you could add an issue template to your repository's CONTRIBUTING.md file, which is linked whenever someone creates a new issue or pull request."