New humanoid action: idle_spawn
The existing spawn action requires spawned humanoids to walk to a tile
after they've spawned, where as this new spawn action doesn't require
this so a humanoid can spawn on a tile and then stand idle on it.

This is how the Grim Reaper spawns in TH, standing idle for a few
seconds before walking to a lava hole. This is why I created this action
and it can play the Grim Reaper's spawn animation and sound effect.

This action's fields:
* point (mandatory):  {x, y, direction}  Where the humanoid will spawn.
* spawn_animation (optional)
* spawn_sound (optional)
* Idle action's fields:  count (optional), loop_callback (optional)