Random VIP Visits
In the manual for TH it says that if you were to refuse the VIP request
to visit too often it would turn up anyway.  This adds that possibility,
fob him off a couple of times and he may just visit you anyway.
Also removed some trailing spaces I found.