Saving Preference
Small change that will allow the game to save the players choice of heat
display colours and for the grant/deny option to allow it to be saved in
the config file.

I feel that with regards to the function comments needing parameter and
return types that these should be done by whoever created them or
someone that knows what they are.  I am only adding something that
allows the choice to be saved between games.
This was first in #98 - trying to squash the commits did not work for me
@TheCycoONE: You still have the option to fire staff that ask for a pay
rise.  All this does is save your chosen preference of what should
happen if you ignore pay demands.  Other preferences can be saved, but
can also be changed within a game.