Pause when the window loses focus Pause when the window loses focus
From issue 1534 on googlecode where the patch was first added a year
ago!
Adds a config option to pause the game when it loses focus - by default
this is off.
Tweaked the original patch from Alan and added a button in the customise
menu
added a button for check for updates option at the same time
amended the no heliport message to make it read better
NOTE: That save game version and afterload function will want updating
as the actual number will depend on when and if this is merged