Remove initial grace period but, reduce negative impact from 0.14 to 0.10
Grace period was introduced as it was reported that CorsixTH became too
cold too quickly. This change started as a report that it was not
sufficiently becoming cold (github issue #74). Rather than make the
period configurable, it was decided to remove the period entirely and
let the user deal with it (github PR #1026).

0.02 / 0.14 = 0.14285714285714285
0.10 is less than that, happiness will decrease less quickly.