I've fixed the faulty Chewbacca fall death animation
The falling part of this animation is faulty with a missing heads layer in frame 30.
So this fix lets the first 21 standing transformation frames be played and then changes the animation to the standard male fall animation by using it as an extra east animation.
This fix also requires the saved game version to be incremented to 83.