Don't open multiple windows at once.
Previously we relied on the THRenderTarget destructor to close the window when
we create a new one. The problem with relying on THRenderTarget::~THRenderTarget
is that it only gets called when the GC collects the object, which doesn't
happen instantly.

Now we call a new THRenderTarget::destroy() method to ensure the window is
destroyed before recreating it.

Fixes GH-113.