Issue 261 Fix For The Continue Button
I've changed how the continue button searches for the latest saved game so that it can search all the directories which are in the saved game directory.

Before this change the continue button code expected there to only be an Autosaves directory and would crash the game if there was other directories in the saved games directory.

New functions
-------------
tree_ctrl.lua:FileTreeNode:
+isDirectory()
+getMostRecentlyModifiedChildFile(file_name_filter):
This new function is now the function responsible for finding the latest saved game for the continue button to load. It is called recursively for all of the directories in the saved games directory.