Made the annual report window do end of year "game won?" checks
This change is required to make my "288_fix" commit always prevent
issue #286 and is also required to make the "end level?" fax window
visible and usable when its shown at the end of a year.

World:onTick() would first display the fax window for the "end level?"
fax by calling OnEndMonth() and then it would call OnEndYear() which
would open the window for the annual report and consequently close the
"end level?" fax window before a player could see and use it. This fax
window's close() function would also unpause the game while the annual
report window was shown.