Fixed World:willObjectsFootprintTileBeWithinItsAllowedRoomIfLocatedAt()
In the final else closure there was a variable scope problem: in both object list search loops the value returned by pairs() was stored in a variable called "object", giving it the same name as the "object" parameter. So I've now renamed these pairs() value variables "additional_objects_name" and "needed_objects_name".

The third problem I've corrected was with the return statement at 1951: within_room would be set to the number value of corridor_object instead of true/false.