Skip action_queue_on_change_position if patient not in queue
When a patient left the queue to be treated in a room while using the
drinks machine, after using the drinks machine the on_change_position
callback would be executed. That callback would enqueue actions to use
the bench again, like the user was still queued. The patient would
then use the bench forever.