Fix tryToFindNearbyPatients for patients using the drinks machine
When a patient in a queue is found and removed from the queue by
tryToFindNearbyPatients, the 'queue' actions is_in_queue flag
must be set to false. Otherwise, when the patient is using the
drinks machine at the moment of the rewrite, the after_use callback
of the drinks machine would create a new walk/use_bench or idle action
for the old queue.