Workaround for patients that fell of a queue
The workaround for patients, that fell of a queue, checks for patients
sitting on a bench while being queued, but are not found for 2 ticks
in the queue. It could happen sometimes, that in the moment before the
tick the patient was dequeued and doing the standing-up anim, which
requires the tick counter. When the patients problem is verfied he
is being sent to the reception.