Allowed rooms to be built over litter objects (all types)
Dealing with issue #147 this commit's gameplay change is based
on what happens in TH when rooms are built over litter.

This commit makes tiles with litter objects on them buildable
while a blueprint is over them and then deletes litter objects
which are in the blueprint when it is confirmed.

New functions added because this commit requires them:
 World.lua:
 +findObjects()
 +findObjectsInArea()
 Utilty.lua:
 +set_get_missing_from()