Fix humanoid empty action queue
When the tickDay function sends a patient to the toilets, there
is a small timeframe (the time required for standing up), where the
tryToFindNearbyPatients function tries to reroute a patient who is
going to the toilet. The finishAction function is then called too
often and all actions from the action queue are consumed, so the error
message pops up.