Fix #370 (zoom glitches)
The `m_pZoomTexture` was being used as a render target even though
it was uninitialized. This sometimes worked out, depending on your
platform, and other times not.

The fix is to initialize the zoom buffer to transparent and then use
SDL_BLENDMODE_BLEND when it is time to flush the zoom buffer.

Fixes #370