Windows Installer: Check if already installed
Installer update:

- Updated welcome.bmp image, simple screenshot from the game looks a
little cleaner (see attached preview).

- Added a check to see if the game is already installed, this is done
via a registry read which pulls out the uninstall path, parses it and
uses that as the currently installed path. The installer then proceeds
with this path for the install. If this data isn't found in the registry
the default directory is used. Tested on Windows 8.0 x64 and Windows 7
x64(TheCycoONE), compiled clean on MakeNSIS v3.0b1.

- Small cleanup of the .nsi file for readability and to remove some old
unused code.