Lua 5.2 compatibility
With this commit the game is fully playable in Lua 5.2.

Several compatibility functions were added or corrected to the luaT C api:
luaT_load - replaces lua_load
luaT_register - replaces lua_register and old luaT_register when second
argument is not null.  (Not assigned to global)
luaT_setfuncs - replaces lua_register when second argument is null
luaT_resume - replaces lua_resume
lua_setfenv - corrections to resulting stack in lua 5.2

luaL_typerror was removed. I replaced it with luaL_argerror with suitable
messages generated for each context (which is what that function did in
5.1).

Other changes to the lua C API required us to change all luaL_reg to
luaL_Reg which is compatible with 5.1 and 5.2.

There were many places in the existing C code which were not using the
compatibility functions already written, they were all changed.

Additionally a few changes had to be made in lua.  In particular:
It is an error in lua 5.2 to pass nil to tonumber.
Compatibility replacements for debug api functions were needed for
persistence.
loadstring is replaced by load in lua 5.2

l_persist_dofile was using all lua_upvalueindex which is
off from luaT_upvalueindex by 1 in lua 5.2 - the environment is stored
at index 1. After several fruitless attempts at finding where the upvalue
was being cross, I decided to change all upvalueindex's to the luaT
compatbility versions.

This as well as the compatibility replacement for debug.getfenv allows
saving games to work.  They can be loaded again in lua 5.1.