Play music tracks in order they appear in MIDI.TXT
The original game played music tracks in the order they appeared in MIDI.TXT. Perhaps this was arbitrary, but perhaps it was intended - either way, when the next track isn't the one I expect it's very jarring!

Prior to this change, MIDI.TXT was loaded, titles extracted, then the playlist was sorted on the titles. This is no less arbitrary than the original order. This change removes the sorting step.