Reimplement THRenderTarget::takeScreenshot.
This creates a surface based on the rendertarget dimensions and a
standard RGB24 format (which SDL_SaveBMP should be able to use directly
without any conversions), rather than trying to retrieve the window
surface of our (potentially hardware-accelerated, which means it doesn't
have a backing surface) window.

Fixes CorsixTH/CorsixTH#739.