Remove bad debugger.lua file check
This file check looked for debugger using io methods, which are not suitable
for determining if the file is in the lua path. This causes the check to fail
when the setup is perfectly fine under numerous scenarios, and on some
platforms prevents running the debugger at all. The dofile will fail with a
missing file error anyway, though it would be clearer if we could replace it.