Eclipse: Created Mars Lua Project & Made Luna Project Lua Only
The Luna project now only has links to folders which contain Lua scripts
because its now meant to be only a Lua project. A seperate C++ project
should be used for CorsixTH's source files because Eclipse's debugger
can only be setup to debug Lua scripts in a Lua project and vice versa
for C++ source files in a C++ project.

No C++ projects will be provided in the Eclipse directory because CMake
can easily be used to generate a compiler ready C++ Eclipse project for
a user's operating system and installed compiler. Its "Eclipse CDT 4"
generator option supports CDT 8.6.0.

The CorsixTH wiki Eclipse tutorial will be updated to include instructions
for generating a Eclipse C++ project with CMake and using it to compile &
debug CorsixTH's source files.

A change to the Lua project's configuration was required for Mars to
make it recognise that this is a Lua project.