Use FindLua included in CMake
The FindLua module included in recent versions of CMake does a better job of
locating lua and using consistent versions for the include directories and
libraries.

In using this module I sacrafice the WITH_LUAJIT flag (not needed), and support
for building on x64 OSX with LuaJIT. It is recommended to build against Lua 5.3.