THMovie optimizations and compatibility improvements
Put the updating of textures on the same thread as rendering so that the GL
context does not need to be shared. Shared contexts are buggy on some hardware.

Use a single hardware texture instead of one texture per frame in the picture
buffer.

Use native height for the texture instead of scaling to fit the window. The
scaling is now done by SDL2 in SDL_RenderCopy rather than sws_scale.
SDL_RenderCopy is capable of scaling the texture on the GPU on most popular
render drivers whereas sws_scale always runs on the CPU.