Do not place walls in unowned plots
When a room is built next to an exterior wall, any windows next to the room's corners are
converted to normal walls to avoid strange looking graphics.
Windows are two tiles wide, but in the case of the border with an unowned plot there may be
only half a window visible. Fixed by not replacing non-wall tiles.