Remove Patient:setdiagDiff()
There is only one use of it. The "self.diagnosis_difficulty" variable is
not set elsewhere, so if expertise fails, the function returns a wrong
result or "nil". The latter would cause a crash on "1 - nil", which
apparently does not happen due to not having reports about it.

In my fix, failure of expertise will always cause a crash, which seems better
than returning a faulty result.