Updated LICENCE.txt for 2016 and expand author list
Authors were pulled from the git logs as well as the copyrights in files.
I encourage anyone missing from the list to update the file with the appropriate
details or contact me with your full name and the name that appears on the
author line of your commits.

Also switches encoding to utf-8 (was latin-1)