Made idle doctors & nurses meandering in rooms answer staff calls
Before this commit doctors and nurses would only answer staff calls when either:
* They started a meander action and weren't in a room.
* They had dealt with a patient and their room had no queuing
patients.

So when a doctor/nurse would be placed in a room they would then ignore staff calls until a patient had used this room.

This would also cause unexpected behaviour when a player had the policy of not allowing staff to leave rooms to answer calls and then disabled it: staff who had become idle while this policy was enabled would not react to it being disabled until after they had dealt with a new patient for the room they were in.