Handle screen shaking in game_ui
Instead of randomly scrolling around the map in an earthquake, include
support for shaking the screen directly in game_ui, where it can be done
without actually moving the window or other hacks and at a consistent rate
regardless of game speed.
At the same time we now exit the tickEarthquake routine immediately if the
game is paused; this avoids retriggering the screen shake.