Add script to build and package libs in MS vcpkg
This commit adds a new script which can be executed on
a Windows machine to build and package the libraries
for CorsixTH. A new folder called vcpkg will be created
in the same folder as the script. The final pacakged
libraries are exported as 7zip files (.7z) ready to
be integrated with CMake.