Remove inline and rename THChunkRenderer methods
Removes the inline keyword from the method declerations. The compiler
can and usually ignores this keyword whilst creating the AST -> Backend
mapping and the optimiser will always do a better job than we can.

Additionally it renames some of the private methods to be more
meaningful and follow the rest of the codebase style