change symbols in zh_CN level introduction text to workaround game hang
The text layout code in the game expects a long string to have spaces,
new lines or line breaks to break the line to suite the display size.
However, Chinese texts usually doesn't satisfy this, so they may hang
the code.

Change full-width Chinese symbols in level introduction texts to western
half-width ones add a space, to workaround this problem.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>