Edited removeKeyHandler. Menu shows current hotkeys.
Edited removeKeyHandler() to coincide with addKeyHandler().

Edited "menu.lua" to allow hotkeys to be correctly represented in in-game menus.
I.E. If someone changed the load game hotkey to "u", then in the "File" menu in-game, the "Load" entry would read "(U) LOAD" instead of "(SHIFT+L) LOAD"

Changed language strings that have references to hotkeys to show those hotkeys in the in-game menu. Some languages missing certain entries.
May want to change modifier names according to language, for instance change "CTRL" to "STRG" for German.

Git is confusing. Fixed an appened(?) in ui.lua.