Mark mouse click handled in file tree ctrl
When selecting a file in the tree control the mouse up event was not
marked as handled, causing it to propogate to underlying windows. In the
map editor this resulted in an error if an overlay was enabled because
the map would be replaced before the event would be handled.