Eliminate static scope from C++
The static keyword is heavily overloaded in C++, so the standard
committee has long recommended using annonymous namespaces instead for
indicating that code only applies to the current translation unit.

As an advantage, anonymous namespaces can also apply to types and
typedefs.

In the case of headers I removed static without replacing it - static
does not make sense in the context of header files.