Better interface for filesystem.lua
Add documentation to filesystem.lua

Expose fileExists and fileSize functions to lua so that we do not
have to read the contents of an iso file  when we only care if the
file exists.

Doesn't leak into bare lfs calls when checking corruption, so that
function works for ISO as well now.

Explicit function for checking file's existence which doesn't leak
implementation details.