Use modern cmake modules for more vcpkg packages
sdl2-mixer and freetype come with cmake modules in vcpkg. Use them instead of
the old FindXXX scripts for those packages when we use vcpkg.

When we update to cmake 3.10 we can get rid of the old path for Freetype.
SDL2 ships with a cmake module upstream, so we might already be able to drop
the alternative path for that dependency.