Replace set_clip_rect and get_clip_rect with push_clip_rect and pop_clip_rect.
Calls to get_clip_rect were used primarily to restore the previous clip, however
due to scaling by the inverse of the scale factor the returned clip would not
always preserve the original clip. By using a stack of clips we can cleanly
restore the previous clip without the caller needing to manually save it off.