Fix #2181
Use a flag to mark if the 255 entry in a palette is transparent.
Store that flag so it will be applied again on load.

This removes the bug that required sprite sheets to be all fetched
on load as well.

Add check to print warning if transparent flag use is mixed, to catch
bugs that could lead to graphical artifacts.

Add luadoc for loadRaw and loadPalette.