Fix array indexing out of bounds in parcel adjacency calculation
The adjacency matrix calculation, which is reponsible for working
out which plots are should be available for purchasing, looks to be
off by an index, due to the way the pointer is incremented prior to
performing the test. This results in undefined behaviour when
testing the edge of the map tiles. It's probably never been noticed
because there's many tiles per parcel.

I've rewritten this to use some of the existing methods to get map
tiles rather than pointer magic, which might be a tad slower but
as this code is infrequently run, I think the tradeoff for
readability is worth it.