Add support for AddressSanitizer, drop VLD
VLD is no longer maintained so it's removed.

Added a convenience option for enabling AddressSanitizer:
https://github.com/google/sanitizers/wiki/AddressSanitizer

The options provided here work in gcc and llvm. Address sanitizer
is available in MSVC but may use a slightly different flag.

I have not enabled it in the windows dev preset because I cannot test
it.