Fix converting lua userdata to animation
reinterpret_cast and static_cast from a void pointer to any type
other than the original is not allowed in c++, there is no
guarantee that the memory is the same, and as soon as virtual
functions are involved it is not.

This fixes that defect as far as it affected this patch by taking
advantage of our metadata to extract the right type of class.