Add support for AddressSanitizer, drop VLD (#2813)
* Add support for AddressSanitizer, drop VLD

VLD is no longer maintained so it's removed.

Added a convenience option for enabling AddressSanitizer:
https://github.com/google/sanitizers/wiki/AddressSanitizer

The options provided here work in gcc and llvm. Address sanitizer
is available in MSVC but uses a slightly different flag. Undefined Behavior sanitizer in not available for MSVC at this time.

I have not enabled it in the windows dev preset because I cannot test
it.