Refactor drawable draw / hit_test / is_multiple_frame_animation methods (#2808)
* Refactor drawable draw/hit_test/is_multiple_frame_animation

- Replaced the 'drawable' base class function pointers by abstract virtual methods.
- Implemented them in 'animation' and 'sprite_render_list'.
- Added an enumeration to 'animation' to track the kind of animation that is stored.

- Deleted all the functions of the removed function pointers.

* Fix converting lua userdata to animation

reinterpret_cast and static_cast from a void pointer to any type
other than the original is not allowed in c++, there is no
guarantee that the memory is the same, and as soon as virtual
functions are involved it is not.

This fixes that defect as far as it affected this patch by taking
advantage of our metadata to extract the right type of class.

---------

Co-authored-by: Stephen E. Baker <baker.stephen.e@gmail.com>