Fix #2826
A regression introduced in 5ff228ac21caeae6aba876728b9b36da9f0a1f5a.

luaT_testuserdata only works correctly if it is passed the correct
metatable reference, and the implementation we had assumed that was the
current environment table, but in the case of map loading it wasn't.

I wrote a replacement method for determining the class of the
animation_base userdata that depends on the __class_info stored in the
metatable.

There was an additional problem with persistence breaking any new saves:
it was using the void* of link_list rather than the class itself.
link_list was not polymorphic so I added the virtual destructor to fix
that and give it a vtable, then dynamic_cast to ensure the proper
pointer type was persisted.