Fix overflow drawing fonts
When drawing non-wrapped text a width of INT_MAX is passed in to the
regular draw_text_wrapped function, along with a left align. This was
overflowing in the alignment offset calculation even though the end
result should be 0. This raised alarms when undefined behaviour
sanitizer was enabled.

Add a conditional to avoid the overflow.