Add Lua side support for ttf font colour selection
The biggest change here was separating loadFont with an exisiting
sprite_sheet from the shortcut method that loaded a spritesheet as well
as the font, which I now call loadFontAndSpriteSheet.

varargs have been removed from the font loading process as well as
parameter overloading making it much easier to follow the code.

A ttf_col parameter takes a red/green/blue colour table and passes it to
setSheet when loading a language font.

The Main Menu font is turned slight off white as a demonstration.