Fix font cache and lua bind casting
Fixes #2885

Fixing the cache eliminated the problem. The cache being
broken was a regression in my code for allowing font colors.

For correctness and because I wrote the code already this patch also
fixes the chain of pointer casting when the base font class is desired,
the same way animation_base was solved.