Avoid calling members on uninitialized objects
When __depersist is called, in most circumstances the object has not yet
be initialized (i.e. by calling new), it is simply a block of malloc'd
memory the right size.

This patch fixes everywhere that was calling methods or accessing
members without calling placement new first.

I'm using void* and _ud as conventions to help future developers not to
accidentally access the object too early; for lack of better enforcement
in the language.

layers is zero initialized by the constructor which removes the need for
the memset.