Support 8bit RGB and RGBA palette files
CorsixTH already used 8-bit per channel RGBA palettes interally but only
supported loading from the original TH format which was 6-bit RGB.

With this change the format can be specified as 6-bit or 8-bit on load
and the 8-bit load routine detects if the palette file contains an
alpha channel based on the size of the file.

This is to support loading palettes extracted from the smacker videos
which FFMPEG exposes as an 8-bit palette with an alpha channel. It
could also be useful for custom artwork, allowing colour selection from
the full True Color gamut and simpler bitmap conversion.

The implementation is consolidated later in the patch series.