movie_player support for headlines
I used vlc and a screenshot utility to measure the dimensions and
location of the headline on the newspaper portion of the video which
appears to match for all 6 lose movies.

The timing of the newspapaer animation and the length of it until the
end of the video is not consistent so I experimentally determined the
time to reveal the headline just by playing through them and tweaking
the values until they looked good to me. It's possible fine adjustments
could be made here by more sensitive eyes or more accurate means of
measuring.

There is one set of headlines in the original data files which are not
used; the rest match the index of the LOSE# movie, so presumably they
intended to develop an additional videos and either didn't or cut it.

There are rendering artifacts at high resolution from scaling the text
independent of the background that I'm not happy with but are out of
scope for now.

The unicode alignment is a little off in a font dependent way, but it
actually looks better than I expected.