Fix #3046 tooltips
I was under the impression that the renderer logical size defaulted to
the window size.

It seems that is mostly true, but when calling
SDL_RendererGetLogicalSize it returns a 0 rather than the window size.

This breaks a couple places where we ask for the dimensions of the
window.

This will revert the behaviour change to #2991 as well.