Add lose.pl8 and winlevel.pl8 in Bitmap directory
lose.pl8 and winlevel.pl8 are extracted from the last frames of
LOSE#.SMK and WINLEVEL.SMK respetively using FFMPEG. The original theme
hospital drew sprites over the videos using the same palette as the
videos rather than shipping a seprate PAL file with the game like is
done with other sprites.

lose.pl8 is used for Font39V and Font40V. winlevel.pl8 is used for
Letter02V as well as some yet unidentified font.

The palette files are created by dumping the data contained in the
data[1] attribute of the video frame, which is populated by
[smacker.c](https://github.com/FFmpeg/FFmpeg/blob/0fdb5829e38dabea9cbe4073a35b6c6315e7508e/libavcodec/smacker.c#L393)

The sprite_viewer is modified to include these palettes in rotation.