Refactor to use font_options table (#3063)
With a growing number of options, passing as method arguments was
becoming unwieldy.

The table is a bit more forgiving with ttf_color, it can now be
specified as a single rgb or argb number i.e. 0xRRGGBB or 0xAARRGGBB if
desired.

It also supports overriding the font size with ttf_width and ttf_height.