Adds a midi player with play xmi files
The player targets the xmi files shipped with CorsixTH which are always
in GM mode.

The midi api and port can be specified in the CorsixTH config.txt file.

Supports all audio functions: play, stop, pause, resume, setVolume.

Two methods of controlling volume are supported for compatibility - the
master volume sysex control which allows fine grained control but lacks
support in many synthesizers, and a method that adjusts the volume
control code events in the midi stream by a factor which is well
supported.

The midi player support is optional, compiling without it forces the
default SDL based playback.