Add isPlaying and sound handles
Rather than directly using channels to reference sound effects use
channels. This way sounds are less likely to coincidentally have the
same ID which was not suitable for an isPlaying method.

Taking advantage of the new handles, add an isPlaying method.

Convert the research dialog to use this method for checking isPlaying.

In the process I discovered a race condition with the reserved channels
which this PR also addresses.