Sound: add isPlaying, handles, and loops (#3146)
* Add isPlaying, loops and sound handles

Rather than directly using channels to reference sound effects use
channels. This way sounds are less likely to coincidentally have the
same ID which was not suitable for an isPlaying method.

Taking advantage of the new handles, add an isPlaying method.

Add loop count to the playSound arguments with -1 for infinite.

Convert the research dialog to use infinite loops.

In the process I discovered a race condition with the reserved channels
which this PR also addresses.