Fix deadlock on sound
SDL_Mixer holds a lock for performing almost all operations, internally
Mix_LockAudio. This lock is not available to code outside of SDL_Mixer
and while there are ways around it they are hacky at best.

Notably that lock is held while calling
sound_player::on_channel_finished. This could lead to a deadlock if we
are already holding channel_mutex and call a Mix_* operation
(on_channel_finished can't get the channel_mutex and Mix_* can't get the
Mix_AudioLock).

With the current SDL_Mixer API I don't see any way to guarantee that the
handle is not obsolete when the operation is called; however that should
be a rare and relatively benign race condition so for now I'll accept
it.

In case a future developer comes across this, my concern is:

Thread 1 calls play, assigns handle A to channel 1.
Thread 1 calls pause with handle A
After getting the channel from playing_channel_for_handle but before
calling Mix_Paused the channel finishes, on_channel_finished is called
and the channel is released.
Then thread 2 called play, assigns handle B to channel 1.
Pause action happens on channel 1 sound B.

I have not been able to come up with a scenario where it causes any
problems with only one main thread playing/stopping/pausing audio which
is the case today.